+++ usher-port/src/usher_common.cpp
@@ -1,114 +1,27 @@
-#include <fstream>
-#include <algorithm>
-#include <numeric>
-#include <boost/program_options.hpp>
-#include <boost/iostreams/filtering_stream.hpp>
-#include <boost/iostreams/filter/gzip.hpp>
-#include <iostream>
-#include <memory>
-#include <limits>
-#include "boost/filesystem.hpp"
-#include "usher_graph.hpp"
-#include "parsimony.pb.h"
-#include "version.hpp"
+#include "usher_common.hpp"
 
 namespace po = boost::program_options;
 namespace MAT = Mutation_Annotated_Tree;
 
-int main(int argc, char** argv) {
-    //Variables to load command-line options using Boost program_options
-    std::string tree_filename;
-    std::string din_filename;
-    std::string dout_filename;
-    std::string outdir;
-    std::string vcf_filename;
-    uint32_t num_cores = tbb::task_scheduler_init::default_num_threads();
-    uint32_t num_threads;
-    uint32_t max_trees;
-    uint32_t max_uncertainty;
-    uint32_t max_parsimony;;
-    bool sort_before_placement_1 = false;
-    bool sort_before_placement_2 = false;
-    bool sort_before_placement_3 = false;
-    bool reverse_sort = false;
-    bool collapse_tree=false;
-    bool collapse_output_tree=false;
-    bool print_uncondensed_tree = false;
-    bool print_parsimony_scores = false;
-    bool retain_original_branch_len = false;
-    bool no_add = false;
-    bool detailed_clades = false;
-    size_t print_subtrees_size=0;
-    size_t print_subtrees_single=0;
-    po::options_description desc{"Options"};
+/**
+ * Preconditions:
+ * - gone through arguments which are saved in the variables
+ * - MAT is loaded
+ * returns: exit code
+ */
 
-    std::string num_threads_message = "Number of threads to use when possible [DEFAULT uses all available cores, " + std::to_string(num_cores) + " detected on this machine]";
-    desc.add_options()
-    ("vcf,v", po::value<std::string>(&vcf_filename)->required(), "Input VCF file (in uncompressed or gzip-compressed .gz format) [REQUIRED]")
-    ("tree,t", po::value<std::string>(&tree_filename)->default_value(""), "Input tree file")
-    ("outdir,d", po::value<std::string>(&outdir)->default_value("."), "Output directory to dump output and log files [DEFAULT uses current directory]")
-    ("load-mutation-annotated-tree,i", po::value<std::string>(&din_filename)->default_value(""), "Load mutation-annotated tree object")
-    ("save-mutation-annotated-tree,o", po::value<std::string>(&dout_filename)->default_value(""), "Save output mutation-annotated tree object to the specified filename")
-    ("sort-before-placement-1,s", po::bool_switch(&sort_before_placement_1), \
-     "Sort new samples based on computed parsimony score and then number of optimal placements before the actual placement [EXPERIMENTAL].")
-    ("sort-before-placement-2,S", po::bool_switch(&sort_before_placement_2), \
-     "Sort new samples based on the number of optimal placements and then the parsimony score before the actual placement [EXPERIMENTAL].")
-    ("sort-before-placement-3,A", po::bool_switch(&sort_before_placement_3), \
-     "Sort new samples based on the number of ambiguous bases [EXPERIMENTAL].")
-    ("reverse-sort,r", po::bool_switch(&reverse_sort), \
-     "Reverse the sorting order of sorting options (sort-before-placement-1 or sort-before-placement-2) [EXPERIMENTAL]")
-    ("collapse-tree,c", po::bool_switch(&collapse_tree), \
[…]
