+++ usher_version.txt
@@ -1,1 +1,1 @@
-962a3d168ab408fbfe32de5bbd0e07b7b2675611
+67b65aae3dba73420d5a76866894da7bba71c0dc
