+++ usher_version.txt
@@ -1,1 +1,1 @@
-67b65aae3dba73420d5a76866894da7bba71c0dc
+96eb926ae8269a5ece30f5eb3f11805455df8ec9
