### 2021-08-31 #3677 @whedon
[REVIEW]: ShUShER: private browser-based placement of sensitive genome samples on phylogenetic trees
**Submitting author:** @amkram (<a href="http://orcid.org/0000-0003-3630-3209">Alexander Kramer</a>)
**Repository:** <a href="https://github.com/amkram/shusher" target ="_blank">https://github.com/amkram/shusher</a>
**Version:** v0.5.0
