+++ usher_version.txt
@@ -1,1 +1,1 @@
-a81569de638ebb904b43701ee65eb4ab826807b3
+6e059a12f53a432d907760eb29ea8033d7888e73
