+++ usher-port/src/usher.cpp
@@ -11,6 +11,7 @@
 #include "usher_graph.hpp"
 #include "parsimony.pb.h"
 #include "version.hpp"
+#include "usher_common.hpp"
 
 namespace po = boost::program_options;
 namespace MAT = Mutation_Annotated_Tree;
@@ -109,106 +110,11 @@ int main(int argc, char** argv) {
         std::cout << "UShER (v" << PROJECT_VERSION << ")" << std::endl;
     }
 
-    if (print_subtrees_size == 1) {
-        std::cerr << "ERROR: print-subtrees-size should be larger than 1\n";
-        return 1;
-    }
-
-    if (sort_before_placement_1 + sort_before_placement_2 + sort_before_placement_3 > 1) {
-        std::cerr << "ERROR: Can't use two or more of sort-before-placement-1, sort-before-placement-2 and sort-before-placement-3 simultaneously. Please specify only one.\n";
-        return 1;
-    }
-
-    if (sort_before_placement_1 || sort_before_placement_2 || sort_before_placement_3) {
-        std::cerr << "WARNING: Using experimental option ";
-        if (sort_before_placement_1) {
-            std::cerr << "--sort-before-placement-1 (-s)\n";
-        }
-        if (sort_before_placement_2) {
-            std::cerr << "--sort-before-placement-2 (-S)\n";
-        }
-        if (sort_before_placement_3) {
-            std::cerr << "--sort-before-placement-3 (-A)\n";
-        }
-    } else if (reverse_sort) {
-        std::cerr << "ERROR: Can't use reverse-sort without sorting options (sort-before-placement-1 or sort-before-placement-2 or sort-before-placement-3)\n";
-        return 1;
-    }
-
-    if (print_parsimony_scores) {
-        if (max_trees > 1) {
-            std::cerr << "ERROR: cannot use --multiple-placements (-M) and --print_parsimony_scores (-p) options simulaneously.\n";
-            return 1;
-        }
-
-        if (sort_before_placement_1 || sort_before_placement_2 || sort_before_placement_3 ||
-                collapse_tree || collapse_output_tree || print_uncondensed_tree || (print_subtrees_size > 0) || (dout_filename != "")) {
-            fprintf (stderr, "WARNING: --print-parsimony-scores-per-node is set. Will terminate without modifying the original tree.\n");
-        }
-    }
-
-    if (max_trees == 0) {
-        std::cerr << "ERROR: Number of trees specified by --multiple-placements (-M) should be >= 1\n";
-        return 1;
-    }
-    /*
-    if (max_trees > 1) {
-        if (max_trees < 256) {
-            std::cerr << "WARNING: Using experimental option --multiple-placements (-M)\n";
-        }
-        else {
-            std::cerr << "ERROR: Number of trees specified by --multiple-placements (-M) should be <= 255\n";
-            return 1;
-        }
-    }
-    */
-
-    if (no_add && (print_subtrees_size > 0 || print_subtrees_single)) {
-        std::cerr << "ERROR: Sorry, cannot output subtrees when -n/--no-add is specified.\n";
-        return 1;
-    }
-
-    if (retain_original_branch_len) {
-        fprintf(stderr, "Output newick files will retain branch lengths from the input tree (unspecified at branches modified during the placement).\n\n");
-    } else {
-        fprintf(stderr, "Output newick files will have branch lengths equal to the number of mutations of that branch.\n\n");
-    }
-
-    boost::filesystem::path path(outdir);
-    if (!boost::filesystem::exists(path)) {
-        fprintf(stderr, "Creating output directory.\n\n");
-        boost::filesystem::create_directory(path);
-    }
-    
-    outdir = path.generic_string();
-
[…]
