+++ README.md
@@ -10,7 +10,7 @@
   
   [![License: AGPL v3](https://img.shields.io/badge/License-AGPL%20v3-blue.svg)](LICENSE)
   [![Integration Tests](https://github.com/amkram/shusher/actions/workflows/build_and_test.yml/badge.svg)](https://github.com/amkram/shusher/actions/workflows/build_and_test.yml)<!---BEGIN_USHER_BADGE-->
-<a target="_blank" href="https://github.com/yatisht/usher/tree/e33eec1198c708a69ff8aafcb86072a98b8fee7a"><img src="https://img.shields.io/badge/UShER%20Version-commit%20e33eec11-%235e0000"/></a>
+<a target="_blank" href="https://github.com/yatisht/usher/tree/eb7061046fb51d1474ad326945786cfdfeb3cfc4"><img src="https://img.shields.io/badge/UShER%20Version-commit%20eb706104-%235e0000"/></a>
 <!---END_USHER_BADGE-->
  
   | :computer_mouse:	Access ShUShER <a target="_blank" href="https://shusher.gi.ucsc.edu">here</a>! |
