+++ usher-port/src/usher_server.cpp
@@ -10,7 +10,6 @@
 #include <memory>
 #include <limits>
 #include <filesystem>
-#include <utility>
 #include "boost/filesystem.hpp"
 #include "usher_graph.hpp"
 #include "parsimony.pb.h"
@@ -20,10 +19,14 @@
 #include <stdlib.h>
 #include <time.h>
 #include <sys/types.h>
-
+#include <sys/stat.h>
+#include <unistd.h>
+#ifdef _WIN32
+#define stat _stat
+#endif
 namespace po = boost::program_options;
 namespace MAT = Mutation_Annotated_Tree;
-namespace fs = boost::filesystem;
+
 
 
 
@@ -31,20 +34,18 @@ namespace fs = boost::filesystem;
 int main(int argc, char** argv) {
 
     //Variables to load command-line options using Boost program_options
-    std::string arg_dirname;
+    std::string arg_filename;
     std::string MAT_list_filename;
     uint32_t num_cores = tbb::task_scheduler_init::default_num_threads();
     po::options_description desc{"Options"};
     uint32_t sleep_length;
-    uint32_t termination_character;
 
 
     desc.add_options()
 
-    ("arguments,a", po::value<std::string>(&arg_dirname)->required(), "Input argument directory that will contain argument files with arguments for usher [REQUIRED]")
+    ("arguments,a", po::value<std::string>(&arg_filename)->required(), "Input argument file that will contain arguments for usher [REQUIRED]")
     ("list-mutation-annonated-trees,i", po::value<std::string>(&MAT_list_filename)->default_value(""), "File containing list of mutation-annotated tree objects")
     ("sleep-length,s", po::value<uint32_t>(&sleep_length)->default_value(100), "Time in milliseconds that the program waits until checking for input in argument file")
-    ("termination-char,c", po::value<uint32_t>(&termination_character)->default_value(94), "Character that will ddetermine if the argument file is ready to be read. Default is '^'")
     ("help,h", "Print help messages");
 
 
@@ -68,20 +69,20 @@ int main(int argc, char** argv) {
     MAT::Tree loaded_MAT;
     bool loaded_MAT_avail = false;//keep track if loaded_MAT is a new copy of MAT and can be used
     std::string loaded_MAT_name = "";
-    
-    fs::path p = fs::current_path();
-    p/=arg_dirname;
-
-    if (!fs::is_directory(p)) {
-        fprintf(stderr, "ERROR: Argument directory provided is not a directory: %s!\n", arg_dirname.c_str());
-        exit(1);
+    if(!boost::filesystem::exists(arg_filename)) {
+        std::cout << "Arguments file not found" <<std::endl;
+        return 1;
     }
+
+    //std::filesystem::path p = std::filesystem::current_path();
+    //std::filesystem::file_time_type modified_time = std::filesystem::file_time_type();
+    time_t modified_time(0);
     MAT::Tree *curr_tree; //MAT that is used in the iteration
     Timer timer;
     std::unordered_map<std::string, MAT::Tree> MAT_list; //store list of trees
     std::unordered_map<std::string, bool> MAT_list_avail; //stores information on whether the MATs in the list are available for use
     if(MAT_list_filename != "") {
-        if(!fs::exists(MAT_list_filename)) {
+        if(!boost::filesystem::exists(MAT_list_filename)) {
             std::cout << "MAT list file not found" <<std::endl;
             return 1;
         }
@@ -122,337 +123,335 @@ int main(int argc, char** argv) {
             }
         }
 
-        if(fs::is_empty(p)) {
+
+        /*
+        if(std::filesystem::last_write_time(p) == modified_time){
+             fprintf(stderr, "Waiting for more arguments\n\n");
[…]
