+++ usher-port/src/usher_common.cpp
@@ -468,15 +468,13 @@ int usher_common(std::string dout_filename, std::string outdir, uint32_t num_thr
                         }
                         if (num_best > max_uncertainty) {
                             fprintf(stderr, "WARNING: Number of parsimony-optimal placements exceeds maximum allowed value (%u). Ignoring sample %s.\n", max_uncertainty, sample.c_str());
-                        }  else if (best_set_difference <= max_parsimony) {
+                        } else if (best_set_difference > max_parsimony) {
+                            fprintf(stderr, "WARNING: Parsimony score of the most parsimonious placement exceeds the maximum allowed value (%u). Ignoring sample %s.\n", max_parsimony, sample.c_str());
+                        } else {
                             fprintf(stderr, "WARNING: Multiple parsimony-optimal placements found. Placement done without high confidence.\n");
                         }
-                    } 
-
-                    if (best_set_difference > max_parsimony) {
-                        fprintf(stderr, "WARNING: Parsimony score of the most parsimonious placement exceeds the maximum allowed value (%u). Ignoring sample %s.\n", max_parsimony, sample.c_str());
                     }
-                }  else {
+                } else {
                     fprintf(stderr, "Missing sample: %s\t Best parsimony score: %d\tNumber of parsimony-optimal placements: %zu\n", sample.c_str(), \
                             best_set_difference, num_best);
 
