### 2021-09-13 #3677 @bede
Review complete. This wasm'd version of the UShER tool constitutes a broadly useful approach, which I am satisfied represents a 'substantial scholarly effort' on the part of its authors. The software is packaged, easily installed and executed. Automated tests are implemented with GitHub Actions. Documentation is brief but clear and in my opinion perfectly adequate. The software paper is well written and includes appropriate references. Great work @amkram
