### 2021-09-15 #3677 @amkram
Thank you @bede! We appreciate your positive feedback, and we have included the JOSS badge in our README per your suggestion.
